Fix __HAL_RCC_CSI_CALIBRATIONVALUE_ADJUST macro
As you can see for `HAL_GetREVID() <= REV_ID_Y` the wrong register is used. It uses `RCC->HSICFGR` instead of `RCC->CSICFGR`, so HSI is configured instead of CSI but with the CSI value.
This leads in a wrong calibration when microcontroller has a revision id lower or equal `REV_ID_Y`.