Fix DMA overflow in ADCDualModeInterleaved example
With `DMA_MDATAALIGN_WORD` the `HAL_ADCEx_MultiModeStart_DMA`
call expects the length parameter to be in words, not half words.

Without this change, the DMA interrupt overwrites the
`aADCxConvertedValues` buffer located immediately behind the,
DMA buffer, which isn't a problem for this example since it's
immediately overwritten with the correct values in the
`HAL_ADC_ConvCpltCallback`.

However, for those using this reference example this bug is likely
to cause memory corruption in their application.